More codebase refactoring
