--- src/apps/admin/components/TempList.scss
@@ -1,9 +1,0 @@
-
-
-.dropZone{
-  border: 2px solid red;
-  display: flex;
-  flex-direction: column;
-  min-height: 500px;
-  justify-content: space-evenly;
-}
