+++ src/apps/admin/layout/Dashboard.tsx
@@ -1,17 +1,56 @@
+import { createContext, useEffect, useState } from 'react'
+import { useSelector } from 'react-redux'
 import { Outlet } from 'react-router-dom'
+import { selectCurrentJukebox, selectCurrentMembership } from 'src/store'
 import { Sidebar, Topbar } from '../components'
 import './Dashboard.scss'
 
+export interface AdminContextType {
+  role: string
+  jukebox: IJukebox | null
+}
+
+export const AdminContext = createContext<AdminContextType>({
+  role: 'user',
+  jukebox: null,
+})
+
 export const Dashboard = () => {
+  const currentJukebox = useSelector(selectCurrentJukebox)
+  const currentMembership = useSelector(selectCurrentMembership)
+
+  const [currentContext, setCurrentContext] = useState({
+    role: 'admin',
+    jukebox: currentJukebox,
+  })
+
+  useEffect(() => {
+    if (currentMembership && currentMembership.is_admin) {
+      setCurrentContext((prev) => ({
+        ...prev,
+        role: 'admin',
+      }))
+    } else {
+      setCurrentContext((prev) => ({
+        ...prev,
+        role: 'member',
+      }))
+    }
+  }, [currentMembership])
+
   return (
-    <div className="dashboard">
-      <div className="dashboard__sidebar">
-        <Sidebar />
+    <AdminContext.Provider value={currentContext}>
+      <div className="dashboard">
+        <div className="dashboard__sidebar">
+          <Sidebar />
+        </div>
+        <div className="dashboard__main">
+          <Topbar />
+          <div className='dashboard__main__outlet'>
+            <Outlet />
+          </div>
+        </div>
       </div>
-      <div className="dashboard__main">
-        <Topbar />
-        <Outlet />
-      </div>
-    </div>
+    </AdminContext.Provider>
   )
 }
