--- src/apps/members/components/Topbar.scss
@@ -1,116 +1,0 @@
-$size: 4rem;
-
-.topbar {
-  background-color: var(--color-surface-container);
-  position: relative;
-  width: 100%;
-
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  z-index: 5;
-
-  &__nav-toggle {
-    position: relative;
-    display: none;
-    
-    @include breakpoint('lg') {
-      display: block;
-    }
-
-    #nav-toggle {
-      position: absolute;
-      width: 100%;
-      height: 100%;
-      top: 0;
-      left: 0;
-      opacity: 0;
-    }
-
-    &__button__icon {
-      position: relative;
-      margin-top: 50%;
-      transform: translate(0, -50%);
-
-      &,
-      &::before,
-      &::after {
-        width: 3rem;
-        height: 2px;
-        background-color: white;
-        display: inline-block;
-      }
-
-      &::before,
-      &::after {
-        content: '';
-        position: absolute;
-        left: 0;
-        transition: all 0.2s;
-      }
-      &::before {
-        top: -0.8rem;
-      }
-      &::after {
-        top: 0.8rem;
-      }
-    }
-  }
-
-  &__group-dropdown {
-    --control-width: 225px; // Used in base styles form component
-    
-    select {
-      padding: 6px;
-      border-radius: 500px;
-      background: inherit;
-      border: 1px solid var(--color-outline-variant);
-      color: #b6b1af;
-      letter-spacing: 1px;
-    }
-  }
-
-  &__user-details {
-    display: flex;
-    gap: 4rem;
-  }
-
-  &__notifications {
-    button {
-      border: none;
-      cursor: pointer;
-      background-color: #dad8d7;
-      border-radius: 500px;
-      width: $size;
-      height: $size;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      &:hover {
-        opacity: 0.8;
-      }
-    }
-  }
-
-  &__profile {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    border: none;
-    cursor: pointer;
-    border-radius: 500px;
-
-    &:hover {
-      opacity: 0.8;
-    }
-
-    img {
-      width: $size;
-      height: $size;
-      object-fit: cover;
-      border-radius: 500px;
-      border: $border-thin;
-    }
-  }
-}
