+++ src/apps/members/layout/MemberDashboard.scss
@@ -1,4 +1,4 @@
-.dashboard {
+.member-dashboard {
   --sidebar-width: 31.4rem;
   --dashboard-gutter: 1rem;
 
@@ -14,30 +14,5 @@
     flex: 1;
     flex-direction: column;
     height: 100%;
-    padding: var(--dashboard-gutter);
-  }
-}
-
-@include breakpoint('lg') {
-  .dashboard:has(.topbar #nav-toggle) .dashboard__sidebar {
-    left: -100%;
-  }
-
-  .dashboard:has(.topbar #nav-toggle:checked) .dashboard__sidebar {
-    left: 0%;
-  }
-
-  .dashboard__sidebar {
-    position: fixed;
-    z-index: 100;
-  }
-
-  .dashboard:has(.topbar #nav-toggle:checked) #nav-toggle {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 99;
   }
 }
