--- src/apps/members/pages/Home.scss
@@ -1,4 +1,0 @@
-.home {
-  background-color: var(--background-color);
-  color: var(--color-text);
-}
