+++ src/apps/members/pages/Home.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-import './Home.scss'
 import { Sidebar } from '../components/Sidebar'
 import { Topbar } from '../components/Topbar'
