--- src/apps/members/pages/MusicQueue.scss
@@ -1,26 +1,0 @@
-.music-queue_container h1 {
-  @include font-headline('sm', 600);
-}
-
-.music-queue_container {
-  display: flex;
-  flex-direction: column;
-  gap: 1%;
-}
-
-.next_up_container {
-  display: flex;
-  flex-direction: column;
-  gap: 2vh;
-}
-
-.recently_played_container {
-  margin-top: 2vh;
-  display: flex;
-  flex-direction: column;
-  gap: 2vh;
-}
-
-.queue-button{
-  margin-top: 5vh;
-}
