--- src/apps/members/pages/MusicSearch.scss
@@ -1,54 +1,0 @@
-.music-search-title {
-  @include font-headline('sm');
-  padding-bottom: 2vh;
-}
-
-.music-search-form {
-  color: var(--color-outline);
-  @include font-label('sm', 300);
-  width: 100%;
-}
-
-.music-search-input {
-  border: 1px solid black;
-  border-radius: 5px;
-  opacity: 50%;
-  padding: 1vh 1vh 1vh 1vh;
-  height: 5vh;
-  width: 100%;
-}
-
-.music-search-input:focus {
-  outline: none;
-}
-
-.music-search-row {
-  display: block;
-  height: 2vh;
-  gap: 20px !important;
-}
-
-.music-search-button-container {
-  grid-column: 11 / span 2 !important;
-}
-
-.music-search-button {
-  height: 5vh;
-  outline: none;
-  border: none;
-  border-radius: 50px;
-  background-color: var(--color-primary-contrast);
-  color: var(--color-primary-on);
-  cursor:pointer;
-
-}
-
-.result-container {
-  margin-top: 8vh;
-}
-
-.track-container {
-  display: flex;
-  flex-direction: column;
-  gap: 2vh;
-}
