+++ src/components/track-list/SearchTrackItem.scss
