+++ src/components/track-list/TrackInteractions.tsx
@@ -1,14 +1,19 @@
 import { ThumbDownAltOutlined, ThumbUpAltOutlined } from '@mui/icons-material'
-import './TrackInteractions.scss'
 import { useContext, useEffect } from 'react'
-import { AdminContext } from 'src/apps/admin/pages/MusicQueue'
-import { Network } from 'src/network'
+import './TrackInteractions.scss'
+
 import { useDrag } from 'react-dnd'
 import { useSelector } from 'react-redux'
+import { AdminContext } from 'src/apps/admin'
+import { MoveIcon, RemoveIcon } from 'src/assets/Icons'
+import { Network } from 'src/network'
 import { selectCurrentMembership } from 'src/store'
 
-export const TrackInteractions = (props: { track?: IQueuedTrack, index?:number }) => {
-  const { track, index} = props
+export const TrackInteractions = (props: {
+  track?: IQueuedTrack
+  index?: number
+}) => {
+  const { track, index } = props
 
   const adminStatus = useContext(AdminContext)
   const network = Network.getInstance()
@@ -16,8 +21,7 @@ export const TrackInteractions = (props: { track?: IQueuedTrack, index?:number }
   const currrentMembership = useSelector(selectCurrentMembership)
 
   const removeTrack = () => {
-    if (track !== undefined && adminStatus.jukebox !== null)
-    {
+    if (track !== undefined && adminStatus.jukebox !== null) {
       console.log(track.queue_id)
       network.removeQueuedTrack(adminStatus.jukebox.id, track.queue_id)
     }
@@ -31,37 +35,39 @@ export const TrackInteractions = (props: { track?: IQueuedTrack, index?:number }
     }),
   })
 
-  useEffect(()=>{
+  useEffect(() => {
     console.log(adminStatus.role)
-  },[currrentMembership])
+  }, [currrentMembership])
 
   return (
     <>
-    {adminStatus.role === 'admin' ? (
-      <div className="track-interactivity">
-        <button className="removeButton" onClick={(e) => {e.stopPropagation(); removeTrack();}}>
-          <svg xmlns="http://www.w3.org/2000/svg" width="15" height="16" viewBox="0 0 15 16" fill="none">
-            <path d="M3.50002 2.50014C3.37208 2.50014 3.244 2.54887 3.1465 2.64663L2.1465 3.64662C1.951 3.84212 1.951 4.15866 2.1465 4.35366L5.79299 8.00014L2.1465 11.6466C1.951 11.8421 1.951 12.1587 2.1465 12.3537L3.1465 13.3537C3.342 13.5492 3.65853 13.5492 3.85353 13.3537L7.50002 9.70717L11.1465 13.3537C11.3415 13.5492 11.6585 13.5492 11.8535 13.3537L12.8535 12.3537C13.049 12.1582 13.049 11.8416 12.8535 11.6466L9.20705 8.00014L12.8535 4.35366C13.049 4.15866 13.049 3.84162 12.8535 3.64662L11.8535 2.64663C11.658 2.45112 11.3415 2.45112 11.1465 2.64663L7.50002 6.29311L3.85353 2.64663C3.75578 2.54887 3.62796 2.50014 3.50002 2.50014Z" fill="#E25850"/>
-          </svg>
-        </button>
-        <button className="moveButton" ref={dragRef} >
-          <svg xmlns="http://www.w3.org/2000/svg" width="15" height="16" viewBox="0 0 15 16" fill="none">
-            <path d="M7.49218 1.99321C7.41687 1.99439 7.3428 2.01257 7.27549 2.04639C7.20819 2.08021 7.1494 2.1288 7.10351 2.18852C7.10285 2.1895 7.1022 2.19047 7.10155 2.19145L5.64648 3.64653C5.59849 3.6926 5.56018 3.74779 5.53379 3.80885C5.50739 3.86992 5.49345 3.93563 5.49277 4.00215C5.4921 4.06867 5.5047 4.13466 5.52984 4.19625C5.55499 4.25784 5.59217 4.31379 5.63921 4.36083C5.68625 4.40787 5.7422 4.44505 5.80379 4.47019C5.86538 4.49534 5.93136 4.50794 5.99788 4.50726C6.0644 4.50659 6.13012 4.49265 6.19119 4.46625C6.25225 4.43986 6.30744 4.40155 6.35351 4.35356L6.99999 3.70708V7.50005H3.20702L3.85351 6.85356C3.92576 6.78333 3.97511 6.69291 3.99509 6.59415C4.01507 6.49539 4.00475 6.3929 3.96547 6.30011C3.9262 6.20731 3.85982 6.12855 3.77501 6.07413C3.69021 6.01972 3.59094 5.99219 3.49023 5.99516C3.36033 5.99903 3.23704 6.05332 3.14648 6.14653L1.69042 7.60259C1.62936 7.64912 1.57983 7.70909 1.54568 7.77785C1.51152 7.84661 1.49366 7.9223 1.49347 7.99908C1.49328 8.07585 1.51077 8.15163 1.54459 8.22056C1.5784 8.28948 1.62764 8.3497 1.68847 8.39653C1.68944 8.39719 1.69042 8.39784 1.6914 8.39848L3.14648 9.85356C3.19255 9.90155 3.24773 9.93986 3.3088 9.96625C3.36986 9.99265 3.43558 10.0066 3.5021 10.0073C3.56862 10.0079 3.63461 9.99534 3.69619 9.97019C3.75778 9.94505 3.81374 9.90787 3.86077 9.86083C3.90781 9.81379 3.94499 9.75784 3.97014 9.69625C3.99528 9.63466 4.00789 9.56867 4.00721 9.50215C4.00653 9.43563 3.99259 9.36992 3.9662 9.30885C3.93981 9.24779 3.90149 9.1926 3.85351 9.14653L3.20702 8.50005H6.99999V12.293L6.35351 11.6465C6.30691 11.5986 6.25119 11.5606 6.18963 11.5346C6.12808 11.5086 6.06193 11.4952 5.99511 11.4952C5.89564 11.4952 5.79844 11.5249 5.71594 11.5804C5.63343 11.636 5.56937 11.7149 5.53194 11.807C5.49451 11.8992 5.48542 12.0004 5.50582 12.0978C5.52622 12.1951 5.5752 12.2842 5.64648 12.3536L7.10253 13.8096C7.14907 13.8707 7.20904 13.9202 7.27779 13.9544C7.34655 13.9885 7.42225 14.0064 7.49902 14.0066C7.57579 14.0068 7.65158 13.9893 7.7205 13.9555C7.78943 13.9216 7.84964 13.8724 7.89648 13.8116C7.89713 13.8106 7.89778 13.8096 7.89843 13.8086L9.35351 12.3536C9.40149 12.3075 9.43981 12.2523 9.4662 12.1912C9.49259 12.1302 9.50654 12.0645 9.50721 11.9979C9.50789 11.9314 9.49529 11.8654 9.47014 11.8038C9.445 11.7423 9.40782 11.6863 9.36078 11.6393C9.31374 11.5922 9.25778 11.555 9.1962 11.5299C9.13461 11.5048 9.06862 11.4921 9.0021 11.4928C8.93558 11.4935 8.86986 11.5074 8.8088 11.5338C8.74773 11.5602 8.69255 11.5985 8.64648 11.6465L7.99999 12.293V8.50005H11.793L11.1465 9.14653C11.0985 9.1926 11.0602 9.24779 11.0338 9.30885C11.0074 9.36992 10.9934 9.43563 10.9928 9.50215C10.9921 9.56868 11.0047 9.63466 11.0298 9.69625C11.055 9.75784 11.0922 9.81379 11.1392 9.86083C11.1862 9.90787 11.2422 9.94505 11.3038 9.9702C11.3654 9.99534 11.4314 10.0079 11.4979 10.0073C11.5644 10.0066 11.6301 9.99265 11.6912 9.96625C11.7523 9.93986 11.8074 9.90155 11.8535 9.85356L13.3096 8.39751C13.3706 8.35097 13.4202 8.291 13.4543 8.22225C13.4885 8.15349 13.5063 8.07779 13.5065 8.00102C13.5067 7.92424 13.4892 7.84846 13.4554 7.77953C13.4216 7.71061 13.3723 7.6504 13.3115 7.60356C13.3105 7.60291 13.3096 7.60226 13.3086 7.60161L11.8535 6.14653C11.8069 6.09863 11.7512 6.06056 11.6896 6.03456C11.6281 6.00856 11.5619 5.99517 11.4951 5.99516C11.3956 5.99519 11.2984 6.02488 11.2159 6.08044C11.1334 6.13599 11.0694 6.21489 11.0319 6.30705C10.9945 6.3992 10.9854 6.50043 11.0058 6.59778C11.0262 6.69513 11.0752 6.78419 11.1465 6.85356L11.793 7.50005H7.99999V3.70708L8.64648 4.35356C8.69255 4.40155 8.74773 4.43986 8.8088 4.46625C8.86986 4.49265 8.93558 4.50659 9.0021 4.50726C9.06862 4.50794 9.13461 4.49534 9.19619 4.47019C9.25778 4.44505 9.31374 4.40787 9.36077 4.36083C9.40781 4.31379 9.445 4.25784 9.47014 4.19625C9.49528 4.13466 9.50789 4.06867 9.50721 4.00215C9.50653 3.93563 9.49259 3.86992 9.4662 3.80885C9.43981 3.74779 9.40149 3.6926 9.35351 3.64653L7.8955 2.18852C7.8481 2.12684 7.78696 2.07706 7.71694 2.04316C7.64693 2.00925 7.56997 1.99215 7.49218 1.99321Z" fill="#B6B1AF"/>
-          </svg>
-        </button>
-      </div>
-    ) : (
-    <div className="track-interactivity">
-      <div className="track-interactivity__item track-interactivity__item--likes">
-        <ThumbUpAltOutlined />
-        {track?.interactions.likes ?? 0}
-      </div>
-      <div className="track-interactivity__item track-interactivity__item--dislikes">
-        <ThumbDownAltOutlined />
-        {track?.interactions.dislikes ?? 0}
-      </div>
-    </div>)
-    }
+      {adminStatus.role === 'admin' ? (
+        <div className="track-interactivity">
+          <button
+            className="removeButton"
+            onClick={(e) => {
+              e.stopPropagation()
+              removeTrack()
+            }}
+          >
+            <RemoveIcon />
+          </button>
+          <button className="moveButton" ref={dragRef}>
+            <MoveIcon />
+          </button>
+        </div>
+      ) : (
+        <div className="track-interactivity">
+          <div className="track-interactivity__item track-interactivity__item--likes">
+            <ThumbUpAltOutlined />
+            {track?.interactions.likes ?? 0}
+          </div>
+          <div className="track-interactivity__item track-interactivity__item--dislikes">
+            <ThumbDownAltOutlined />
+            {track?.interactions.dislikes ?? 0}
+          </div>
+        </div>
+      )}
     </>
   )
 }
